refactor(utils): moved some helper function to rpc `utils.rs`
